AIStudyCompanionServer/requirements.txt
flask
torch
torchvision
torchaudio
transformers
openai-whisper
